added all command functionality
all the commands have now had there execution added into struct functions implement by traits. The general rule of thumb is, if the command is used by the server struct, it implements the ServerRunnables trait, and if the command is used by the client struct, it implements the ClientRunnables trait. This ensures each command can be used correctly.